+++ src/components/dlc/platforms/DZBankProRes/DZBankProres.js
@@ -1,0 +1,5 @@
+// Test-Platform
+export async function DZBankProRes(bookingData){
+    console.log('DZBank ProRes Starting')
+    return "ProRes OK"
+}